const prop tests passing

+++ compiler/src/test/scala/edg/compiler/ExprEvaluateTest.scala
@@ -187,20 +187,20 @@ class ExprEvaluateTest extends AnyFlatSpec {
 
   it should "handle if then else" in {
     evalTest.map(
-      ValueExpr.IfThenElse(ValueExpr.Literal(true), ValueExpr.Literal(4.0), ValueExpr.Literal(3))
-    ) should equal(FloatValue(4.0f))
+      ValueExpr.IfThenElse(ValueExpr.Literal(true), ValueExpr.Literal(4), ValueExpr.Literal(3))
+    ) should equal(IntValue(4))
     evalTest.map(
-      ValueExpr.IfThenElse(ValueExpr.Literal(false), ValueExpr.Literal(4.0), ValueExpr.Literal(3))
-    ) should equal(FloatValue(3.0f))
+      ValueExpr.IfThenElse(ValueExpr.Literal(false), ValueExpr.Literal(4), ValueExpr.Literal(3))
+    ) should equal(IntValue(3))
 
     import edg.expr.expr.BinaryExpr.Op
     evalTest.map(
       ValueExpr.IfThenElse(ValueExpr.BinOp(Op.AND, ValueExpr.Literal(true), ValueExpr.Literal(true)),
-      ValueExpr.Literal(4.0), ValueExpr.Literal(3))
-    ) should equal(FloatValue(4.0f))
+      ValueExpr.Literal(4), ValueExpr.Literal(3))
+    ) should equal(IntValue(4))
     evalTest.map(
       ValueExpr.IfThenElse(ValueExpr.BinOp(Op.AND, ValueExpr.Literal(true), ValueExpr.Literal(false)),
-        ValueExpr.Literal(4.0), ValueExpr.Literal(3))
-    ) should equal(FloatValue(3.0f))
+        ValueExpr.Literal(4), ValueExpr.Literal(3))
+    ) should equal(IntValue(3))
   }
 }
